requirements.txt
asgiref==3.2.10
cairocffi==0.9.0
CairoSVG==2.4.2
cffi==1.14.1
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1
django-tables2==2.3.1
djangorestframework==3.11.1
html5lib==1.1
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
xlrd==1.2.0
